Move '/alive', '/now', and '/version' endpoints to their own InfoController in each service
