+++ bitwarden_license/src/Sso/Controllers/HomeController.cs
@@ -5,7 +5,6 @@ using IdentityServer4.Services;
 using System.Threading.Tasks;
 using Bit.Sso.Models;
 using System.Diagnostics;
-using System.Reflection;
 using Microsoft.AspNetCore.Diagnostics;
 
 namespace Bit.Sso.Controllers
@@ -17,23 +16,6 @@ namespace Bit.Sso.Controllers
         public HomeController(IIdentityServerInteractionService interaction)
         {
             _interaction = interaction;
-        }
-
-        [HttpGet("~/alive")]
-        [HttpGet("~/now")]
-        [AllowAnonymous]
-        public DateTime GetAlive()
-        {
-            return DateTime.UtcNow;
-        }
-        
-        [HttpGet("~/version")]
-        [AllowAnonymous]
-        public JsonResult GetVersion()
-        {
-            var assembly = Assembly.GetExecutingAssembly();
-            var fileVersionInfo = FileVersionInfo.GetVersionInfo(assembly.Location);
-            return Json(fileVersionInfo.ProductVersion);
         }
 
         [Route("~/Error")]
