+++ bitwarden_license/src/Sso/Controllers/InfoController.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using Microsoft.AspNetCore.Mvc;
+using Bit.Core.Utilities;
+
+namespace Bit.Admin.Controllers
+{
+    public class InfoController : Controller
+    {
+        [HttpGet("~/alive")]
+        [HttpGet("~/now")]
+        public DateTime GetAlive()
+        {
+            return DateTime.UtcNow;
+        }
+        
+        [HttpGet("~/version")]
+        public JsonResult GetVersion()
+        {
+            return Json(CoreHelpers.GetVersion());
+        }
+    }
+}
