--- src/Identity/Controllers/MiscController.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Diagnostics;
-using System.Reflection;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Bit.Identity.Controllers
-{
-    public class MiscController : Controller
-    {
-        public MiscController() { }
-
-        [HttpGet("~/alive")]
-        [HttpGet("~/now")]
-        public DateTime GetAlive()
-        {
-            return DateTime.UtcNow;
-        }
-
-        [HttpGet("~/version")]
-        public JsonResult GetVersion()
-        {
-            var assembly = Assembly.GetExecutingAssembly();
-            var fileVersionInfo = FileVersionInfo.GetVersionInfo(assembly.Location);
-            return Json(fileVersionInfo.ProductVersion);
-        }
-    }
-}
