+++ src/Notifications/Controllers/SendController.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Diagnostics;
-using System.IO;
-using System.Reflection;
+﻿using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using Bit.Core.Utilities;
@@ -19,23 +16,6 @@ namespace Bit.Notifications
         public SendController(IHubContext<NotificationsHub> hubContext)
         {
             _hubContext = hubContext;
-        }
-
-        [HttpGet("~/alive")]
-        [HttpGet("~/now")]
-        [AllowAnonymous]
-        public DateTime GetAlive()
-        {
-            return DateTime.UtcNow;
-        }
-        
-        [HttpGet("~/version")]
-        [AllowAnonymous]
-        public JsonResult GetVersion()
-        {
-            var assembly = Assembly.GetExecutingAssembly();
-            var fileVersionInfo = FileVersionInfo.GetVersionInfo(assembly.Location);
-            return Json(fileVersionInfo.ProductVersion);
         }
 
         [HttpPost("~/send")]
